move validateAndFixTree to utils and use in setTree, don't use checkTree
