+++ modules/stores/tree.js
@@ -18,7 +18,7 @@ import {
   getNewValueForFieldOp
 } from "../utils/ruleUtils";
 import {deepEqual, defaultValue, applyToJS} from "../utils/stuff";
-import {validateValue} from "../utils/validation";
+import {validateValue, validateAndFixTree} from "../utils/validation";
 import omit from "lodash/omit";
 import mapValues from "lodash/mapValues";
 
@@ -714,7 +714,8 @@ export default (config) => {
 
     switch (action.type) {
     case constants.SET_TREE: {
-      set.tree = action.tree;
+      const validatedTree = validateAndFixTree(action.tree, state.tree, action.config || config, config);
+      set.tree = validatedTree;
       break;
     }
